Extract FeedCachePolicy into its own file
