+++ EssentialFeed/EssentialFeed/Feed Cache/FeedCachePolicy.swift
@@ -1,0 +1,24 @@
+//
+//  FeedCachePolicy.swift
+//  EssentialFeed
+//
+//  Created by Jules on 19/02/2025.
+//
+
+import Foundation
+
+class FeedCachePolicy {
+    
+    private static let calendar = Calendar(identifier: .gregorian)
+
+    private static var maxCacheAgeInDays : Int {
+        return 7
+    }
+    
+    static func validate(_ timestamp: Date, against date: Date) -> Bool {
+        guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else {
+            return false
+        }
+        return date < maxCacheAge
+    }
+}
